requirements.txt
Django>=4.2,<5
djangorestframework>=3.14.0
requests>=2.28.1
celery>=5.4
redis>=5.0
